requirements.txt
discord.py
praw
peewee
flask
flask_login
oauthlib
redis
gunicorn
gevent
config